[round-7] 결제 성공/실패 이벤트 추가 (#27)
* refactor: 결제 성공 이벤트 추가

* refactor: 결제 실패 이벤트 추가
