+++ apps/commerce-api/src/main/java/com/loopers/domain/payment/PaymentEventPublisher.java
@@ -1,0 +1,7 @@
+package com.loopers.domain.payment;
+
+public interface PaymentEventPublisher {
+    void publish(PaymentEvent.Success event);
+
+    void publish(PaymentEvent.Fail event);
+}
