+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/payment/PaymentV1Controller.java
@@ -3,6 +3,8 @@ package com.loopers.interfaces.api.payment;
 import com.loopers.application.payment.PaymentCriteria;
 import com.loopers.application.payment.PaymentFacade;
 import com.loopers.application.payment.PaymentResult;
+import com.loopers.domain.payment.PaymentCommand;
+import com.loopers.domain.payment.PaymentService;
 import com.loopers.interfaces.api.ApiResponse;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -18,15 +20,14 @@ import org.springframework.web.bind.annotation.RestController;
 public class PaymentV1Controller implements PaymentV1ApiSpec {
 
     private final PaymentFacade paymentFacade;
+    private final PaymentService paymentService;
 
     @Override
     @PostMapping("/callback")
     public ApiResponse<?> callback(@RequestBody PaymentV1Dto.CallbackRequest request) {
         switch (request.status()) {
-            case SUCCESS -> paymentFacade.success(
-                    new PaymentCriteria.Success(request.transactionKey(), UUID.fromString(request.orderId())));
-            case FAILED -> paymentFacade.fail(
-                    new PaymentCriteria.Fail(request.transactionKey(), UUID.fromString(request.orderId()), request.reason()));
+            case SUCCESS -> paymentService.success(new PaymentCommand.Success(request.transactionKey()));
+            case FAILED -> paymentService.fail(new PaymentCommand.Fail(request.transactionKey(), request.reason()));
         }
         return ApiResponse.success();
     }
