+++ apps/commerce-api/src/main/java/com/loopers/interfaces/scheduler/PaymentScheduler.java
@@ -1,8 +1,7 @@
 package com.loopers.interfaces.scheduler;
 
-import com.loopers.application.payment.PaymentCriteria;
-import com.loopers.application.payment.PaymentFacade;
 import com.loopers.domain.payment.Payment;
+import com.loopers.domain.payment.PaymentCommand;
 import com.loopers.domain.payment.PaymentInfo;
 import com.loopers.domain.payment.PaymentService;
 import java.util.List;
@@ -15,17 +14,14 @@ import org.springframework.stereotype.Component;
 public class PaymentScheduler {
 
     private final PaymentService paymentService;
-    private final PaymentFacade paymentFacade;
 
     @Scheduled(cron = "0 0/1 * * * ?")
     public void syncPayment() {
         List<PaymentInfo.Transaction> transactionInfos = paymentService.getUnsyncedPendingPayments();
         for (PaymentInfo.Transaction info : transactionInfos) {
             switch (info.status()) {
-                case Payment.Status.COMPLETED ->
-                        paymentFacade.success(new PaymentCriteria.Success(info.transactionKey(), info.orderId()));
-                case Payment.Status.FAILED ->
-                        paymentFacade.fail(new PaymentCriteria.Fail(info.transactionKey(), info.orderId(), info.reason()));
+                case Payment.Status.COMPLETED -> paymentService.success(new PaymentCommand.Success(info.transactionKey()));
+                case Payment.Status.FAILED -> paymentService.fail(new PaymentCommand.Fail(info.transactionKey(), info.reason()));
             }
         }
     }
