+++ apps/commerce-api/src/test/java/com/loopers/domain/payment/PaymentServiceIntegrationTest.java
@@ -3,18 +3,24 @@ package com.loopers.domain.payment;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.any;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 import com.loopers.infrastructure.payment.gateway.LoopersPaymentGateway;
+import com.loopers.utils.DatabaseCleanUp;
 import java.math.BigDecimal;
 import java.util.UUID;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.context.event.RecordApplicationEvents;
 
 @SpringBootTest
+@RecordApplicationEvents
 class PaymentServiceIntegrationTest {
 
     @Autowired
@@ -22,7 +28,16 @@ class PaymentServiceIntegrationTest {
     @Autowired
     private PaymentRepository paymentRepository;
     @MockitoBean
+    private PaymentEventPublisher paymentEventPublisher;
+    @MockitoBean
     private LoopersPaymentGateway paymentGateway;
+    @Autowired
+    private DatabaseCleanUp databaseCleanUp;
+
+    @AfterEach
+    void tearDown() {
+        databaseCleanUp.truncateAllTables();
+    }
 
     @Nested
     @DisplayName("결제 요청 시,")
@@ -38,6 +53,40 @@ class PaymentServiceIntegrationTest {
             PaymentInfo paymentInfo = paymentService.pay(command);
 
             assertThat(paymentInfo.transactionKey()).isEqualTo("transactionKey123");
+        }
+    }
+
+    @Nested
+    @DisplayName("결제 성공 시,")
+    class Success {
+
+        @Test
+        @DisplayName("결제 성공 이벤트를 발행한다.")
+        void publishPaymentSuccessEvent() {
+            Payment payment = Payment.of(new PaymentCommand.Pay(new BigDecimal("1000"), UUID.randomUUID(), "SAMSUNG", "1234-1234-1234-1234"));
+            payment.successRequest("TX-KEY");
+            paymentRepository.save(payment);
+
+            paymentService.success(new PaymentCommand.Success(payment.getTransactionKey()));
+
+            verify(paymentEventPublisher, times(1)).publish(new PaymentEvent.Success(payment.getTransactionKey(), payment.getOrderId()));
+        }
+    }
+
+    @Nested
+    @DisplayName("결제 실패 시,")
+    class Fail {
+
+        @Test
+        @DisplayName("결제 실패 이벤트를 발행한다.")
+        void publishPaymentFailEvent() {
+            Payment payment = Payment.of(new PaymentCommand.Pay(new BigDecimal("1000"), UUID.randomUUID(), "SAMSUNG", "1234-1234-1234-1234"));
+            payment.successRequest("TX-KEY");
+            paymentRepository.save(payment);
+
+            paymentService.fail(new PaymentCommand.Fail(payment.getTransactionKey(), "결제 실패"));
+
+            verify(paymentEventPublisher, times(1)).publish(new PaymentEvent.Fail(payment.getTransactionKey(), payment.getOrderId(), "결제 실패"));
         }
     }
 }
